FSPT-1013: properly separate tasklist section and submission statuses
Initially submission statuses directly reflected the status for each
section in the tasklist, not started, in progress and completed.

Now that the submission workflow involves more steps we've started to
separate the two enums but they continue to half depend on each other.

Separate these out and update type signatures to make it clear in the
code which is calculated the status of a section and whats part of the
submission workflow status.