FSPT-985: Error handlers understand app context
Now our concept of two distinct services is coming together, our error
handlers and templates should understand the correct service context
and at least show the right service name in the header to avoid user
confusion.

This change extends the helper function which decides which service
desk URL to pass to the templates to also decide which base templates
the error template should extend, which would show the correct service
name and can be extended in the future to show more service context.