FSPT-1026: Add sent_for_certification_by to SubmissionHelper
Adds a property to the SubmissionHelper to get the user who submitted
the report for certification, and show this in the view_locked_report
view.

Also to guard against a future state where we have multiple Submission
Events for sending a report for certification (eg. in the event of a
report being declined and then re-sent for certification) we also add
make sure all submission events are sorted by created_at_utc for
consistency and ease in grabbing the latest one we need.

We also add default sorting to the SubmissionEvent relationship
on the Submission model, but adding it explicitly on the the Submission
Helper property so developers know the ordering is important.