FSPT-1026: use the submission event user for certify notification email
This swaps out passing the in-memory "current_user()" to send the
notification email to certifiers letting them know who submitted the
report.

By using the user we've recorded made that submission event we can be
confident we can send this email any time with a deterministic result -
i.e if we need to retry it it doesn't rely on the users session.

Uses the property added that filters submission events.