FSPT-1026: redirect logic for is you have bookmarked the submission page
If for example a certifier refects a submission we probably want to be
showing people the tasklist, if we're not in our expected scenario send
users back to the list reports page for now (which should avoid an
infite redirect in the case its just a mistake on our end -- could think
about the obersavability for that)

Calls out that we should be checking submission IDs belong the org and
grant you are loading the page for in case someone is able to get access
to submission IDs from other users.

Also favours using generic submit form as per review comments.