FSPT-1028: Update confirmation page backlink
In the course of FSPT-1028 we realised the backlink on the 'report sent
for certification' success page backlink was taking you back to the
tasklist, which didn't make sense in this context. This changes the back
link to take you back to the `route_to_submission` which encapsulates
routing logic.