FSPT-1018: submission events v2
Updates the submission events table to prepare for "reducing" streams of
submission events to allow workflows and state to go both backwards and
forwards.

This will allow us to model submissions being signed off, declined,
submitted and marked as overdue.

We'll also be able to streamline how individual forms/ sections are
tracked with events so we're no longer deleting events when a user goes
back and changes if its completed or not but can use the properties
enabled by this change.