FSPT-1018: add events helper library
Adds a small lib to represent submission events and their properties.

By processed events in a reduced stream like this, either grouped by a
form on the submission or the sumbission itself we should be able to
flexibility model the workflow and audit requirements of submissions:
- awaiting sign off (if certification required)
- being signed off (if certification required)
- being declined (if certification required)
- being submitted

Originally tried using pydantic but as the type validation didn't feel
useful and the overhead in syntax to make it do something it didn't seem
like it wanted to do did't feel worth it - uses native dataclasses
instead.

Note shallow copy is used when extracting metadata interfaces from
submission event models as `asdict` creates a deep copy which means
you'll end up with two ORM references which doesn't feel right.