FSPT-1018: connect new events strategy
Remove the current "enum" method of checking for the existence of events
and instead check the existing "stream" or list of submission events and
their properties (using the reducer helper).

This allows us to replace the code that previously removed submission
events and move to an append only approach that should work well now
that we'll have more involved workflows (i.e a decline both impacting
the submission and the submissions forms).

To verify this all works, the existing functionality should be
unchanged!