FSPT-1018: add remaining submission workflow events
Adds the remaining events to support the full submission workflow for
reports.

As events record who saved them and when we can extract a fairly
thorough history of the submission through metadata which are available
to the helpers to be shown on pages and used in emails.

Some of the business logic for if certification steps and approvals
should be mandatory are waiting on a flag on collections which can turn
certification off or on.