BAU: VS Code settings for formatting HTML on save
VS Code's default HTML format on save does not follow our desired
HTML formatting rules, which are applied with prettier through Make
commands and used in our pre-commit hooks.

This change adds all the configurations passed to the make-frontend
commands to the prettierrc config file and updates the VS Code
settings.json to use these settings as the formatter for HTML files,
and default to tab width 4 for the other filetypes we format with Prettier.