BAU: use the functioning tasklist page until certifier screens are done
We prematurely updating the routing logic in AGF to a work-in-progress
flow while the test environment is being used for UR and demos. This
feels like a clear need for us to round out an initial pass at feature
flags to let us safely merge code as its put together.

For now revert routing to this page so that everything continues to work
as expected for test environment users.