FSPT-1017: Add auth check to certify/decline/submit
There are instances (eg certify, submit) where our auth handlers on routes
allow access to users who should not be able to do certain actions on the
page, so even though those actions are protected in the template
we should still guard against them in case users are able to trigger
form submissions etc.

This adds auth checks to the certify/decline/submit methods in the auth
helper to ensure these can only proceed if the user has the correct
permissions.

This doesn't apply to test submissions which should still be able to
be submitted, and will only apply to collections that require
certification.