FSPT-1019: use playwright browser to generate reports
Adds an endpoint that brings in the playwright SDK to spin up chromium
and generate a PDF on the backend.

This works for now and these projects are well maintained. We also
considered weasyprint and wkhtmltopdf but weasyprint was not working
nicely with the design system CSS (page was displayed differently to
expected without much time to correct this, additionally some CSS styles not
supported).

We should review this as the method for generating PDFs as there are
some drawbacks:
- generally can spin up and close down within 200ms but simultaenous
  requests in a sync framework quickly slow down
- including a headless browser in our little app deployment doesn't feel
  ideal - a library thats able to run and directly convert to pdf bytes
feels cleaner and more scalable