BAU: Fix 'view submission' styling & breadcrumb
Fixes a couple of things noticed during a demo:
- live submissions were showing the submitted user's email in the
breadcrumb (as it is for test submissions) instead of the grant
recipient org
- 'Submitted' status doesn't have a tag and was showing as raw text on
the the view submission page, as previously we only saw this status
in the list reports table where it would be styled with govuk table cell
styling. The fix here doesn't look _great_ but we've talked about
updating the layout of this page to match the 'view_locked_report' page,
in which case it would become redundant as that doesn't display the
status when submitted (or would display it elsewhere).