FSPT-1019: buildpacks w/ deb dependencies
Avoid managing system dependencies after the buildpack layers are
configured (the playwright script to install dependencies uses the root
user which is blocked by buildpacks, we instead delegate to another
buildpack to make sure we have the required deps).

Uses an inline buildpack to make sure that a cut back headless version
of chromium is built.