BAU: Send certify confirmation/decline emails to all users
We previously sent the decline certification and certification/submit
success email notifications only to those users who had sent the report
for certification and had declined or certified it. In discussion with
product it was decided that we should default to sending these emails
to all grant recipient data provider and certifier users, rather than
just the individuals who had triggered those submission events. This is
something we will monitor and can roll back if it gets too spammy.