Set grant code during setup
Allows users creating grants in Deliver grant funding to set its unique code
during the grant setup flow.

This is not easily changeable after the fact, as it's used to generate
submission references and that could lead to old references looking/feeling
disconnected.

If the grant code needs to change this will be a manual developer action.