BAU: tidy up test/ preview banner
Make the banners full screen so that they feel part of the structure
rather than part of the page.

Also change so to use a similar style to the govuk button inverse which
is for use on banners/ darker backgrounds - this feels like it "ties"
the button in with the banner but happy to tweak or revert if this
doesn't feel right.