FIX: request commit handling being triggered on failed requests
The default behaviour for the Flask middleware stack is to not call the
`after_request` hook for failed responses - this meant that the auto commit
after request deorator and handler would avoid committing failures.

However since that was checked in we've added `@app.errorhandlers` to
show specific well formatted pages for 4xx and 5xx failures. This is a
better user experience but the response object that is returned from
these methods is not considered to be the failured response and actually
does trigger `after_request` which means that if exceptions are raised
the auto commit after request decorator will still run.

For now check if the response is successful before ever committing - we
might want to extend this and reconsider where it sits in the Flask
middleware lifecycle.

Tests verifying this behaviour need to turn `TESTING` to False as
otherwise exceptions are immediately raised up and stop the test.