FSPT-1061: make access flows work without reporting dates
We expect baseline reports won't have meaningful reporting dates for
grant recipients. Rely on the name being a proxy for reporting dates if
thats meaningful or being the appropriate thing to identify the report
otherwise.

Tries to standardise access ("Dates to be confirmed") and deliver ("Not
set") terminology so this can be replaced simply if we decide to tweak
it.