FSPT-1061: provide is test data to Notify emails
Tell all AGF transactional emails if they're test data or not. This also
stops the Notify interface from raising if there is missing data as that
should be handled upstream.

Not done as part of this but if something is considered absolutely
required (i.e it is the address you'll be sending the email to it
should be required in the method interface)