FSPT-1061: directly sso challenge if logging in with internal email
Instead of redirecting to a different context directly upgrade the login
request to SSO if an internal email is used. If the user is requesting
a link to sign in to AGF and don't have a previously set next URL set
them to access AGF after this - if they don't have any AGF permissions a
no permissions page will show and they have a link back to DGF from
there.

To complete this we should remove the code to remove all non
platform admin permissions on successful platform admin login so that any
AGF scenarios that have been set up don't get cleared each time.