Re-validate submission on submit
When completing a submission, some questions may be validated based on answers
to earlier questions. It's possible to go back and change previous answers at
any time, which may then mean existing answers to future questions become
invalid. We don't detect this when re-answering old questions, so currently if
this happens nothing will detect it and it would be possible to submit with
invalid data.

We don't have time to address this properly right now, so this patch adds a
snapshot final check at submission time to revaldiate all of the data in the
form. This will ensure data submitted is all valid together, but doesn't give
the user a nice experience when this happens as we haven't had time to design a
good experience around it. We'll show an error summary on the submission
tasklist page and list all forms+question names+answers that have errored, to do
a very rough attempt to help the user self-serve. But we'll direct them to
support if they are stuck.