FSPT-921: remove add another entry
While working on add another we included a link to remove entries that
you no longer needed and built out the backend logic to support this. We
never connected the frontend form runner pages and the ability to remove
a given entry.

Add an add another context page to confirm if the entry should be
removed and then connect the backend interface to remove data at a given
entry.

In the future we probably want to allow you to remove add another
entries before you've added all of the information for that entry
(currently "Change" and "Remove" will only show when an entry is
complete) but that thinking should probably be done alongside if we
should show a check your entry answers for more involved add another
question groups.