BAU: Fix submitting preview submissions
We recently moved sending confirmation emails from routes to submission
helper methods for consistency, but this would break when trying to
submit a preview submission (ie when testing a form) as there's no grant
recipient in a preview submission and we shouldn't be sending confirmation
emails for previews.

This just adds a check to ensure emails only send if it's not a preview
submission (ie. it's Test or Live).