BAU: Consolidate confirmation pages
We had two near-identical confirmation pages for when a submission is
submitted to MHCLG, one for signed off & submitted (when a report
requires certification) and just submitted (when a report doesn't
require it). These had near-identical route handlers as well.

This brings these two pages and routes together and updates the template
to handle the small content differences if a report requires
certification or not (including the relevant feedback url).

Also fixes a URL path to have hyphens rather than underscores.