BAU: Add interstitial page & hook up for certify flow
As part of pre-launch checks, we've decided to add an interstitial page
to the submission flow. Adding an extra page to this flow
where the user has to again confirm their intent to submit adds a little
more friction and should avoid accidental submissions from the
view_locked_report or tasklist pages.

This commit adds the new page and hooks it up for the submission flow
for reports that require certification. The next commit will hook it up
for reports that don't require certification and submit directly from
the tasklist.