BAU: add another remove link for complete and incomplete entries
This will allow form users to remove work in progress answers in an add
another group.

Previously you had to fill in every answer before it removed the call to
action to finish and allow you to "Change" and "Remove".

This anticipated have two or three questions in a question group but the
reality of add another for upcoming forms are ~10-20 questions.

With the current behaviour you'd have to fill in "dummy" answers to get
to the point where you could remove it - if you undestood that was the
implication.

Allow removing any entry thats been started.