BAU: Add interstitial page to no certification submission flow
Now that we have interstitial page set up to confirm submission, this
hooks it up to the submission flow when a report doesn't require
certification.

We still need to ensure that the entire report is validated again before
it's submitted, which was previously done by the submit() method called
from the tasklist, so we call the full form validation instead so we can
still append errors to the tasklist page and block the user from
continuing until the errors are fixed.