BAU: Catch full-form validation errors
With the new interstitial page, reports that don't require certification
could circumvent full-form navigation by amending the URL and jumping
to the new interstitial confirm submission page when the form has
validation errors which are normally caught by clicking the 'continue to
submit' button on the tasklist. Not handled, these would throw full
validation errors and cause a 500.

As a temporary solution until we handle full-form validation better,
we now catch these errors coming from the 'submit' method called in the
route, redirect back to the tasklist and render the same error message
as the user would receive if they'd clicked the 'contine to submit'
button from the tasklist.

This commit also adds some additional test coverage for the form state
redirects at the top of the route to ensure users can only ever land on
this page when the form is in the correct state.