Add individual data provider page
When a report is open, it's unlikely we'll ever want to bulk add more data
providers and instead we're responding to individual support requests. So let's
change the page from a bulk add to an individual add.

When setting up new users, we often want to send an onboarding email as well, so
let's include that option on the page.

Whether the best email to send is the 'report is ready to complete' email or a
new 'you've been added to an existing report' email is slightly up for grabs -
but currently we send the report ready to complete email manually.