FIX: treat empty add another lists as unanswered questions
We don't currently support optional questions but when we check for
completeness in an add another context (a list of answers) we're checking that
all of the entries have all of their questions completely answered.

If there are no entries no missing or invalid answers will be flagged, which
should be considered the base case if no answers have been provided at all.