Refactor submission event data
Stop storing static submission event data (ie attributes that are implicit in
the event type and never vary). We only need to store data that can vary by
individual event instance, which in our case so far is just the 'declined
reason' of a certifier declining a report.

All other attributes can be inferred based on the event type, so let's do that.