FSPT-1049: move overdue out of the submission status
This moves the overdue state out of the status calculation and into its
own property. This allows all of the existing workflows to continue to
work even if a submission is overdue but gives us the option to visually
differentiate overdue submissions.

Overdue is calculated by checking if the submission deadline has passed
for the collection, and that the submission isn't in a terminal
(submitted) state.

This lets us tweak what it means for a given submission (test, preview,
live) to be overdue but means we can check if a collection is the past
uniformly (for example there is no submission if the user hasn't started
one).

Updates the status tag to visually differentiate overdue submissions.